Moves files into directories
