+++ index.js
@@ -3,7 +3,7 @@ const _isEmpty = require('lodash/isEmpty'),
   _includes = require('lodash/includes'),
   passport = require('passport'),
   flash = require('express-flash'),
-  { compileLoginPage } = require('./login'),
+  { compileLoginPage } = require('./services/login'),
   {
     getAuthUrl,
     getPathOrBase,
@@ -11,7 +11,7 @@ const _isEmpty = require('lodash/isEmpty'),
     deserializeUser,
     getProviders
   } = require('./utils'),
-  sessionStore = require('./session-store'),
+  sessionStore = require('./services/session-store'),
   { addAuthRoutes, createStrategy } = require('./strategies'),
   { AUTH_LEVELS } = require('./constants'),
   { withAuthLevel } = require('./services/auth'),
