+++ services/login.test.js
@@ -4,7 +4,7 @@ const _startCase = require('lodash/startCase'),
   handlebars = require('handlebars'),
   filename = __filename.split('/').pop().split('.').shift(),
   lib = require(`./${filename}`),
-  utils = require('./utils');
+  utils = require('../utils');
 
 describe(_startCase(filename), function () {
   describe('compileLoginPage', function () {
